mission/code/requirements.txt
smolagents[toolkit]
torch
transformers
Pillow
opencv-python
requests
markdownify
pygame
mutagen
pyarrow
